feat(kickstart): Add base Kickstart file
Add the base Kickstart file, which configures a base installation of the OS,
and installs the daemons necessary for use with the Microsft Hyper-V
hypervisor.

This however, does not include the root credentials - installation will halt
and prompt the user for the root password. This is to be automated through
Packer using boot parameters (see GH-2).

This installation has been tested on CentOS.

Associated issue: GH-1